+++ api/mrts.py
@@ -1,5 +1,5 @@
 from flask import *
-from model.database import execute_query
+from model.mrts import get_mrt_names
 
 # 使用 Blueprint 創建路由
 mrts_bp = Blueprint('mrts', __name__)
@@ -18,25 +18,3 @@ def get_mrts():
             "message": "伺服器內部錯誤"
         }
         return jsonify(error_response), 500
-
-
-def get_mrt_names():
-    query = """
-        SELECT a.mrt 
-        FROM attractions a
-        GROUP BY a.mrt
-        ORDER BY COUNT(*) DESC, a.mrt
-    """
-
-    result = execute_query(query)
-
-    mrts = []
-    count = 0
-    for row in result:
-        if row[0] is not None:
-            mrts.append(row[0])
-            count += 1
-            if count >= 40:
-                break
-
-    return {"data": mrts}
